move ContentType to right package
